products/requirements.txt
Django>=2,<3
djangorestframework==3.11.0
markdown
python-decouple==3.1
django-filter==2.1.0
django-events-sourcing>=0.2
nameko>=2.12.0
django-nameko-standalone
psycopg2-binary
